daily/20200106/example_spreadsheet/requirements.txt
gspread
python-dotenv
google-auth-oauthlib
dictknife
